# MINOR: rewrite zipWithIndex by normal foreach to refrain unnecessary …