@chia7712 2020-08-26
> Do you have any data on the improvement this achieves?

@ijuma the JMH result is attached. please take a look.