@chia7712 2020-09-23
```
Build / JDK 15 / kafka.network.DynamicConnectionQuotaTest.testDynamicListenerConnectionCreationRateQuota
```
it is traced by #9301

```
org.apache.kafka.streams.processor.internals.StateDirectoryTest.shouldLogStateDirCleanerMessage
```

https://issues.apache.org/jira/browse/KAFKA-10514