@ijuma 2020-09-24
Unrelated flaky failures:

```
Build / JDK 15 / kafka.network.ConnectionQuotasTest.testListenerConnectionRateLimitWhenActualRateBelowLimit | 5 sec | 1
-- | -- | --
Build / JDK 8 / org.apache.kafka.streams.integration.EosBetaUpgradeIntegrationTest.shouldUpgradeFromEosAlphaToEosBeta[true] | 2 min 21 sec | 1
Build / JDK 8 / org.apache.kafka.streams.integration.StandbyTaskEOSIntegrationTest.shouldWipeOutStandbyStateDirectoryIfCheckpointIsMissing[exactly_once]
```